use msys on windows, override install command
The "make install" command requires the "install" command on windows.
This is most easily provided by Alien::MSYS (via the alien_msys directive).

Also the makefile assumes that USER is set and (incorrectly) that the
user has a corresponding group with the same name.  We can get around
this by overriding INSTALL in the "make install" line